backend/requirements.txt
djangorestframework,
markdown,
django-filter,
chromadb,
PyMuPDF,
python-docx,
dotenv,
mistralai,
mysqlclient,
sentence-transformers